+++ db/migrate/20210403164118_create_bids.rb
@@ -1,7 +1,7 @@
 class CreateBids < ActiveRecord::Migration[6.1]
   def change
     create_table :bids do |t|
-      t.integer :amount, default: 0
+      t.float :amount, default: 0
       t.belongs_to :user
       t.belongs_to :item
       t.timestamps
